refactor: builder
