+++ code/compiler/builder/common/util.FolderSize.mts
