+++ code/compiler/builder/index.mts
@@ -1,3 +1,4 @@
-export { fs } from './common/index.mjs';
 export { Builder } from './Builder/index.mjs';
 export { Count } from './Count.mjs';
+export { Paths } from './Paths.mjs';
+export { fs } from './common/index.mjs';
