+++ code/compiler/builder/op/Vite.mts
@@ -1,7 +1,7 @@
 import type { ConfigEnv, UserConfigFn } from 'vite';
 import { build } from 'vite';
 
-import { Config } from '../Config.mjs';
+import { Config } from '../../Config.mjs';
 import { Paths } from '../Paths.mjs';
 import { Template } from '../Template.mjs';
 import { Util, fs, type t } from '../common/index.mjs';
