+++ script.ts/common/index.mts
@@ -1,5 +1,5 @@
 import type * as t from './types.mjs';
 
-export type { t };
-export { Util } from './util.mjs';
 export * from './libs.mjs';
+export { Util } from './util.mjs';
+export { type t };
